[FIX] Fixed the retrun type of a fm related function and adjusted the masai test to a changed fm index member
git-svn-id: http://svn.seqan.de/seqan/trunk@14393 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ core/include/seqan/index/index_fm_prefix_sum_table.h
@@ -86,16 +86,17 @@ struct GetValue<PrefixSumTable<TChar, TSpec> const>
 // ----------------------------------------------------------------------------
 // Metafunction clear
 // ----------------------------------------------------------------------------
+// TODO(singer): change the type to be dependent on a template argument
 template <typename TChar, typename TSpec>
 struct Fibre<PrefixSumTable<TChar, TSpec>, FibreEntries>
 {
-    typedef String<unsigned long> Type;
+    typedef String<unsigned long long> Type;
 };
 
 template <typename TChar, typename TSpec>
 struct Fibre<PrefixSumTable<TChar, TSpec> const, FibreEntries>
 {
-    typedef String<unsigned long> const Type;
+    typedef String<unsigned long long> const Type;
 };
 
 // ----------------------------------------------------------------------------
